# server/docs/swagger.json
{
    "swagger": "2.0",
    "info": {
        "description": "apis List",
        "title": "Go1.16 Gin框架-文档自动生成",
        "contact": {},
        "version": "1.0.0"
    },
    "basePath": "/",
    "paths": {
...
        "/znFlight/createZnFlight": {
            "post": {
                "security": [
                    {
                        "ApiKeyAuth": []
                    }
                ],
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "ZnFlight"
                ],
                "summary": "创建ZnFlight",
                "parameters": [
                    {
                        "description": "创建ZnFlight",
                        "name": "data",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/autocode.ZnFlight"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "{\"success\":true,\"data\":{},\"msg\":\"获取成功\"}",
                        "schema": {
                            "type": "string"
                        }
                    }
                }
            }
        },
        "/znFlight/deleteZnFlight": {
            "delete": {
                "security": [
                    {
                        "ApiKeyAuth": []
                    }
                ],
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "ZnFlight"
                ],
                "summary": "删除ZnFlight",
                "parameters": [
                    {
                        "description": "删除ZnFlight",
                        "name": "data",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/autocode.ZnFlight"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "{\"success\":true,\"data\":{},\"msg\":\"删除成功\"}",
                        "schema": {
                            "type": "string"
                        }
                    }
                }
            }
        },
        "/znFlight/deleteZnFlightByIds": {
            "delete": {
                "security": [
                    {
                        "ApiKeyAuth": []
                    }
                ],
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "ZnFlight"
                ],
                "summary": "批量删除ZnFlight",
                "parameters": [
                    {
                        "description": "批量删除ZnFlight",
                        "name": "data",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/request.IdsReq"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "{\"success\":true,\"data\":{},\"msg\":\"批量删除成功\"}",
                        "schema": {
                            "type": "string"
                        }
                    }
                }
            }
        },
        "/znFlight/findZnFlight": {
            "get": {
                "security": [
                    {
                        "ApiKeyAuth": []
                    }
                ],
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "ZnFlight"
                ],
                "summary": "用id查询ZnFlight",
                "parameters": [
                    {
                        "type": "string",
                        "description": "创建时间",
                        "name": "createdAt",
                        "in": "query"
                    },
                    {
                        "type": "integer",
                        "description": "主键ID",
                        "name": "id",
                        "in": "query"
                    },
                    {
                        "type": "string",
                        "name": "inarrive",
                        "in": "query"
                    },
                    {
                        "type": "string",
                        "name": "inautc",
                        "in": "query"
                    },
                    {
                        "type": "string",
                        "name": "incode",
                        "in": "query"
                    },
                    {
                        "type": "string",
                        "name": "indate",
                        "in": "query"
                    },
                    {
                        "type": "string",
                        "name": "iniata",
                        "in": "query"
                    },
                    {
                        "type": "string",
                        "name": "intime",
                        "in": "query"
                    },
                    {
                        "type": "string",
                        "name": "outarrive",
                        "in": "query"
                    },
                    {
                        "type": "string",
                        "name": "outautc",
                        "in": "query"
                    },
                    {
                        "type": "string",
                        "name": "outcode",
                        "in": "query"
                    },
                    {
                        "type": "string",
                        "name": "outdate",
                        "in": "query"
                    },
                    {
                        "type": "string",
                        "name": "outiata",
                        "in": "query"
                    },
                    {
                        "type": "string",
                        "name": "outtime",
                        "in": "query"
                    },
                    {
                        "type": "string",
                        "description": "更新时间",
                        "name": "updatedAt",
                        "in": "query"
                    }
                ],
                "responses": {
                    "200": {
                        "description": "{\"success\":true,\"data\":{},\"msg\":\"查询成功\"}",
                        "schema": {
                            "type": "string"
                        }
                    }
                }
            }
        },
        "/znFlight/getZnFlightList": {
            "get": {
                "security": [
                    {
                        "ApiKeyAuth": []
                    }
                ],
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "ZnFlight"
                ],
                "summary": "分页获取ZnFlight列表",
                "parameters": [
                    {
                        "type": "string",
                        "description": "创建时间",
                        "name": "createdAt",
                        "in": "query"
                    },
                    {
                        "type": "integer",
                        "description": "主键ID",
                        "name": "id",
                        "in": "query"
                    },
                    {
                        "type": "string",
                        "name": "inarrive",
                        "in": "query"
                    },
                    {
                        "type": "string",
                        "name": "inautc",
                        "in": "query"
                    },
                    {
                        "type": "string",
                        "name": "incode",
                        "in": "query"
                    },
                    {
                        "type": "string",
                        "name": "indate",
                        "in": "query"
                    },
                    {
                        "type": "string",
                        "name": "iniata",
                        "in": "query"
                    },
                    {
                        "type": "string",
                        "name": "intime",
                        "in": "query"
                    },
                    {
                        "type": "string",
                        "name": "outarrive",
                        "in": "query"
                    },
                    {
                        "type": "string",
                        "name": "outautc",
                        "in": "query"
                    },
                    {
                        "type": "string",
                        "name": "outcode",
                        "in": "query"
                    },
                    {
                        "type": "string",
                        "name": "outdate",
                        "in": "query"
                    },
                    {
                        "type": "string",
                        "name": "outiata",
                        "in": "query"
                    },
                    {
                        "type": "string",
                        "name": "outtime",
                        "in": "query"
                    },
                    {
                        "type": "integer",
                        "description": "页码",
                        "name": "page",
                        "in": "query"
                    },
                    {
                        "type": "integer",
                        "description": "每页大小",
                        "name": "pageSize",
                        "in": "query"
                    },
                    {
                        "type": "string",
                        "description": "更新时间",
                        "name": "updatedAt",
                        "in": "query"
                    }
                ],
                "responses": {
                    "200": {
                        "description": "{\"success\":true,\"data\":{},\"msg\":\"获取成功\"}",
                        "schema": {
                            "type": "string"
                        }
                    }
                }
            }
        },
        "/znFlight/updateZnFlight": {
            "put": {
                "security": [
                    {
                        "ApiKeyAuth": []
                    }
                ],
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "ZnFlight"
                ],
                "summary": "更新ZnFlight",
                "parameters": [
                    {
                        "description": "更新ZnFlight",
                        "name": "data",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/autocode.ZnFlight"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "{\"success\":true,\"data\":{},\"msg\":\"更新成功\"}",
                        "schema": {
                            "type": "string"
                        }
                    }
                }
            }
        }
    },
    "definitions": {
...
        "autocode.ZnFlight": {
            "type": "object",
            "properties": {
                "createdAt": {
                    "description": "创建时间",
                    "type": "string"
                },
                "id": {
                    "description": "主键ID",
                    "type": "integer"
                },
                "inarrive": {
                    "type": "string"
                },
                "inautc": {
                    "type": "string"
                },
                "incode": {
                    "type": "string"
                },
                "indate": {
                    "type": "string"
                },
                "iniata": {
                    "type": "string"
                },
                "intime": {
                    "type": "string"
                },
                "outarrive": {
                    "type": "string"
                },
                "outautc": {
                    "type": "string"
                },
                "outcode": {
                    "type": "string"
                },
                "outdate": {
                    "type": "string"
                },
                "outiata": {
                    "type": "string"
                },
                "outtime": {
                    "type": "string"
                },
                "updatedAt": {
                    "description": "更新时间",
                    "type": "string"
                }
            }
        },
...
        "request.IdsReq": {
            "type": "object",
            "properties": {
                "ids": {
                    "type": "array",
                    "items": {
                        "type": "integer"
                    }
                }
            }
        },
...
    },
    "securityDefinitions": {
        "ApiKeyAuth": {
            "type": "apiKey",
            "name": "x-token",
            "in": "header"
        }
    }
}